==> 03 - followup - predicting sudden gains.Rmd <==
---
title: "sudden gains followup"
author: "ryan schubert"
date: "1/11/2022"
output: html_document
---

```{r setup, include=FALSE}
library(dplyr)
library(ranger)
library(gbm)
library(caret)
library(tidyr)
library(data.table)
library(plotrix)
library(ggplot2)
library(viridis)
library(lme4)
library(lmerTest)
library(caret)
library(ROCR)
"%&%" = function(a,b) paste0(a,b)
set.seed(1234)
```


```{r}
pcl.dir<-"/Users/ryanschubert/Dropbox/Ryan's stuff/rush/ML project/datasets/3 Week IOP/"
data.dir<-"/Users/ryanschubert/Dropbox/Ryan's stuff/rush/sudden gains/"
out.dir<-"/Users/ryanschubert/Dropbox/Ryan's stuff/rush/sudden gains/"
```

There are several follow ups to the sudden gains

```{r}
rawData<-fread(pcl.dir %&% "IOP_DATA_SET_6-12-2020 with Manual Corrections copy.csv" ,na=c("-"))
sg<-fread(out.dir %&% "complete_cases_sudden_gains.csv")
table(sg$sg_crit123,sg$sg_freq_byperson)
```

Question 1 is how does a sudden gain affect the maintenance of gains - eg do people with a sudden gain show better maintenance at three months compared to those who do not?

```{r}

pcl_time<-rawData %>% select(PAT_ID,contains("PCL_"))  %>% select(-PCL_SCORE_POST) %>% pivot_longer(!PAT_ID)

long_data<-sg %>% select(PAT_ID,sg_crit123) %>% inner_join(pcl_time,by=c("PAT_ID")) 
plot_data<-long_data %>%
  group_by(sg_crit123,name) %>%
  summarise(mean_score=mean(value,na.rm=T),
            se_score=std.error(value,na.rm=T)) %>%
  mutate(name=factor(name,levels=c("PCL_SCORE","PCL_DAY2","PCL_DAY3","PCL_DAY5","PCL_DAY6","PCL_DAY8", "PCL_DAY10","PCL_DAY11","PCL_DAY13","PCL_DAY14","PCL_SCORE_3MON", "PCL_SCORE_6MON","PCL_SCORE_12MON")),
         sg_crit123=if_else(sg_crit123 == 1 , "SG Present","No SG"))
tmp<-ggplot(plot_data,aes(x=name,y=mean_score)) +
  geom_line(aes(group=sg_crit123,colour=sg_crit123)) +
  geom_errorbar(aes(ymin=mean_score-se_score,ymax=mean_score+se_score,colour=sg_crit123)) +
  geom_point(aes(shape=sg_crit123)) +
  theme_bw() +
  theme(axis.text.x=element_text(angle=45,hjust=1),
        axis.title.x = element_blank()) +
  labs(colour="SG Present",shape="SG Present",y="Average PCL-5")
ggsave(out.dir %&% "sg_PCL_scores_over_time_with_followup.png",dpi=300,height=4.9,width=7,plot=tmp)

# how many people with an sg even have followup data

pcl_time<-rawData %>% select(PAT_ID,contains("PCL_"))  %>% select(-PCL_SCORE_POST)
long_data<-sg %>% select(PAT_ID,sg_crit123) %>% inner_join(pcl_time,by=c("PAT_ID")) 

count_na<-function(vec){
  sum(is.na(vec)) 
}

table(long_data$sg_crit123)
tapply(long_data$PCL_SCORE_3MON,long_data$sg_crit123,summary)
tapply(long_data$PCL_SCORE_6MON,long_data$sg_crit123,summary)
tapply(long_data$PCL_SCORE_12MON,long_data$sg_crit123,summary)
table(sg$sg_crit123)
```

Question 2 is when are people having sudden gains?

```{r}
table(sg$sg_session_n)
timing<-sg %>% select(sg_session_n) %>%
  mutate(session=case_when(
    sg_session_n == 1 ~ "Day 2",
    sg_session_n == 2 ~ "Day 3",
    sg_session_n == 3 ~ "Day 5",
    sg_session_n == 4 ~ "Day 6",
    sg_session_n == 5 ~ "Day 8",
    sg_session_n == 6 ~ "Day 10",
    sg_session_n == 7 ~ "Day 11"
  )) %>%
  filter(!is.na(timing)) %>%
  mutate(session=factor(session,levels=c("Day 2","Day 3","Day 5","Day 6","Day 8","Day 10","Day 11")))
plotTiming<-ggplot(timing,aes(x=session)) +
  geom_bar() +
  theme_bw()
ggsave(out.dir %&% "Histogram_when_sudden_gains.png",dpi=300,height=4.9,width=7,plot=plotTiming)

```


Question 3 is can we predict if a person will have a sudden gain or not? If yes can we predict when?

```{r}
labels<-sg %>% select(PAT_ID,sg_crit123)
vars<-c("PAT_ID",
        "PCL" %&% 1:20,
        "PTCI" %&% 1:33,
        "PHQ" %&% 1:9,
        "NSI" %&% 1:22,
        "AUDITC" %&% 1:3,
        "COHORT_TYPE",
        "gender","AGE_AT_COHORT_START","FirstRace","Ethnicity")

colnames(rawData)
variables<-rawData %>% select(one_of(vars)) %>%
  mutate(FirstRace=if_else(FirstRace=="White","White","Nonwhite"),
         Ethnicity=if_else(Ethnicity=="Hispanic or Latino","H","NH"))
modeling_data<-inner_join(labels,variables,by=c("PAT_ID"))
modeling_data$FirstRace
# modeling_data$PAT_ID[duplicated(modeling_data$PAT_ID)]
```

```{r}
modeling_data<-modeling_data %>% select(-PAT_ID)
modeling_data<-mutate(modeling_data,
                      sg_crit123=if_else(sg_crit123==1,"SG","nSG"),
                      sg_crit123=as.factor(sg_crit123))
```

```{r}
k<-10
folds<-createFolds(modeling_data$sg_crit123,k=k,list=F)
table(folds,modeling_data$sg_crit123)

gbm_tunegrid<-expand.grid(interaction.depth = c(1,2),n.trees = c(500),shrinkage = c(0.001,0.01,0.1),n.minobsinnode = c(1,3,5))
rf_grid<-expand.grid(mtry = 2,splitrule = c("gini","extratrees"),min.node.size = c(1, 3, 5))

rf_roc<-rep(NA,k)
gbm_roc<-rep(NA,k)
for (i in 1:k){
  holdin<-modeling_data[i!=folds,]
  holdout<-modeling_data[i==folds,]
  
  gbm_ctrl <- trainControl(method="cv", number=5, returnResamp="all",classProbs=TRUE,summaryFunction=twoClassSummary)
  gbm_model <- caret::train(sg_crit123 ~ .,data=holdin,method = "gbm", trControl = gbm_ctrl,verbose=F,metric='ROC',distribution="bernoulli",tuneGrid = gbm_tunegrid)
  
  rf_ctrl <- trainControl(method="cv", number=5, returnResamp = "all",classProbs=TRUE)
  rf_model<- caret::train(sg_crit123 ~ .,data =holdin,method = "ranger",verbose=F,trControl = rf_ctrl,metric="ROC",distribution="bernoulli",tuneGrid = rf_grid)
  
  #get performance
  preds<-predict(rf_model,holdout,type="prob")[,1]
  predObj<-ROCR::prediction(preds,holdout$sg_crit123)
  perf<-ROCR::performance(predObj,"auc")
  rf_roc[i]<-perf@y.values[[1]]
  
  preds<-predict(gbm_model,holdout,type="prob")[,1]
  predObj<-ROCR::prediction(preds,holdout$sg_crit123)
  perf<-ROCR::performance(predObj,"auc")
  gbm_roc[i]<-perf@y.values[[1]]
  
}
print(1-rf_roc)
mean(1-rf_roc)
print(1-gbm_roc)
mean(1-gbm_roc)
```



Question 4 is why is there such a large variability in start and end point for people without sudden gain
