==> 11_ML_pred_2_week_ITP.Rmd <==
---
title: "11 ML predict two week"
author: "ryan schubert"
date: "2/25/2022"
output: html_document
---

```{r setup, include=FALSE}
library(dplyr)
library(ranger)
library(gbm)
library(caret)
library(tidyr)
library(data.table)
library(plotrix)
library(ggplot2)
library(viridis)
library(lme4)
library(lmerTest)
library(caret)
library(ROCR)
library(naniar)
library(suddengains)
"%&%" = function(a,b) paste0(a,b)
set.seed(1234)
```

```{r cars}
out.dir<-"/Users/ryanschubert/Dropbox/Ryan's stuff/rush/sudden gains/"
dir<-"/Users/ryanschubert/Dropbox/Ryan's stuff/rush/2week/"
list.files(dir)
rawData<-fread(dir %&% "Two_week_with_CAPS_2_25_22_extended.csv" ,na=c("-"))
colnames(rawData)[grepl("pcl5_score",colnames(rawData),ignore.case=T)]
PCL_Data<-rawData %>% select(mrn,contains("pcl5_score"),contains("mst"),Age,Sex,cohort_type_base) %>%
  select(-contains("followup")) %>%
  mutate(MSTYN=case_when(
    !is.na(mst1_base) & mst1_base == 1 ~ 1,
    !is.na(mst2_base) & mst2_base == 1 ~ 1,
    !is.na(mst3_base) & mst3_base == 1 ~ 1,
    !is.na(cohort_type_base) & cohort_type_base == 2 ~ 1,
    T ~ 0
  ))
duplicated_mrn<-PCL_Data$mrn[duplicated(PCL_Data$mrn)] %>% unique()

PCL_Data<-PCL_Data %>% filter(!(mrn %in% duplicated_mrn))
PCL_with_outcome<-PCL_Data %>% filter(!is.na(pcl5_score_post))
```


```{r}
count_not_na<-function(vec){
  sum(!is.na(vec))
}

count_not_na_windowise<-function(vec,window=5,n_sess=5){
  mod<-(length(vec) - window + 1)
  windowise<-rep(NA,mod)
  for (i in 1:mod){
    windowise[i]<-(sum(!is.na(vec[i:(i+window-1)])) >=n_sess)
  }
  return(windowise)
}

summarise_windowise<-function(vec){
  sum(count_not_na_windowise(vec)) >=1
}
summarise_windowise(c(1,1,1,NA,1,1,1,1,1))

available_data<-PCL_Data %>%
  select(starts_with("pcl")) %>%
  apply(1,summarise_windowise)
sufficient_data<-PCL_Data %>% 
  mutate(not_na=available_data) %>%
  filter(available_data==T,
         !is.na(pcl5_score_post),
         !is.na(pcl5_score_pastmonth_base))
vis_miss(sufficient_data)
```


```{r}
cutoff<-define_crit1_cutoff(data_sd=sufficient_data$pcl5_score_pastmonth_base,reliability=0.84)
vars<-colnames(PCL_Data)[2:11]
byperson <- create_byperson(data = sufficient_data,
                    sg_crit1_cutoff=cutoff$reliable_change_value,
                    id_var_name = "mrn",
                    tx_start_var_name = "pcl5_score_pastmonth_base",
                    tx_end_var_name = "pcl5_score_post",
                    sg_var_list=vars,
                    identify_sg_1to2=T,
                    sg_measure_name = "PCL",
                    identify = "sg")
byperson$sg_crit123
table(byperson$sg_reversal)
byperson <- byperson %>% mutate(stable_sg=if_else(sg_crit123 == 1 & sg_reversal != 1, 1,0))
table(byperson$stable_sg)
table(byperson$stable_sg)/nrow(byperson)
table(byperson$sg_freq_byperson)
cutoff$reliable_change_value
summary(byperson$sg_session_n)


```


```{r}
outcomes<-byperson %>% select(mrn,stable_sg)
colnames(rawData)[grepl("mie",colnames(rawData),ignore.case=T)]
vars<-c("mrn",
        "cohort_type_base",
        "ptci" %&% 1:33 %&% "_base",
        "pcl" %&% 1:20 %&% "_pastmonth_base",
        "phq" %&% 1:9 %&% "_base",
        "nsi" %&% 1:22 %&% "_base",
        "auditc" %&% 1:3 %&% "_new_base",
        "auditc" %&% 1:3 %&% "_old_base",
        "mies" %&% 1:11 %&% "_base",
        "mst" %&% 1:3 %&% "_base",
        "Sex",
        "Age",
        "FirstRace (group)",
        "Ethnicity")
modeling_data<-rawData %>% select(one_of(vars)) %>% 
  right_join(outcomes,by=c("mrn")) %>%
  mutate(`FirstRace (group)`=if_else(`FirstRace (group)`=="White","White","Nonwhite"),
         Ethnicity=if_else(Ethnicity=="Hispanic or Latino","H","NH"),
         stable_sg=if_else(stable_sg==1,"SG","NoSG"),
         stable_sg=as.factor(stable_sg),
         auditc1=if_else(!is.na(auditc1_old_base),auditc1_old_base,auditc1_new_base),
         auditc2=if_else(!is.na(auditc2_old_base),auditc2_old_base,auditc2_new_base),
         auditc3=if_else(!is.na(auditc3_old_base),auditc3_old_base,auditc3_new_base),
         MSTYN=case_when(
            !is.na(mst1_base) & mst1_base == 1 ~ 1,
            !is.na(mst2_base) & mst2_base == 1 ~ 1,
            !is.na(mst3_base) & mst3_base == 1 ~ 1,
            !is.na(cohort_type_base) & cohort_type_base == 2 ~ 1,
            T ~ 0
          )) %>%
  select(-mrn,-contains("_old_"),-contains("_new"),-mst1_base,-mst2_base,-mst3_base,-cohort_type_base) %>%
  filter(complete.cases(.))
vis_miss(modeling_data)
fwrite(modeling_data,out.dir %&% "Two_week_modeling_data.csv")
```

```{r}

k<-10
folds<-createFolds(modeling_data$stable_sg,k=k,list=F)
table(folds,modeling_data$stable_sg)

svm_tunegrid<-expand.grid(degree=c(1:3),scale=c(0.01,0.001, 0.0001),C = c(0.1,1,10,50))
rf_grid<-expand.grid(mtry = 2,splitrule = c("gini","extratrees"),min.node.size = c(1, 3, 5))
gbm_tunegrid<-expand.grid(interaction.depth = c(1,2),n.trees = c(500),shrinkage = c(0.001,0.01,0.1),n.minobsinnode = c(1,3,5))

rf_roc<-rep(NA,k)
svm_roc<-rep(NA,k)
gbm_roc<-rep(NA,k)

rf_sens<-rep(NA,k)
svm_sens<-rep(NA,k)
gbm_sens<-rep(NA,k)

rf_spec<-rep(NA,k)
svm_spec<-rep(NA,k)
gbm_spec<-rep(NA,k)
for (i in 1:k){
  holdin<-modeling_data[i!=folds,]
  holdout<-modeling_data[i==folds,]
  
    gbm_ctrl <- trainControl(method="cv", number=5, returnResamp="all",classProbs=TRUE,summaryFunction=twoClassSummary)
  gbm_model <- caret::train(stable_sg ~ .,data=holdin,method = "gbm", trControl = gbm_ctrl,verbose=F,metric='ROC',distribution="bernoulli",tuneGrid = gbm_tunegrid)
  
  svm_ctrl <- trainControl(method="cv", number=5, returnResamp="all",classProbs=TRUE,summaryFunction=twoClassSummary)
  svm_model <- caret::train(stable_sg ~ .,data=holdin,method = "svmPoly", trControl = svm_ctrl,verbose=F,metric='ROC',distribution="bernoulli",tuneGrid = svm_tunegrid)
  
  rf_ctrl <- trainControl(method="cv", number=5, returnResamp = "all",classProbs=TRUE)
  rf_model<- caret::train(stable_sg ~ .,data =holdin,method = "ranger",verbose=F,trControl = rf_ctrl,metric="ROC",distribution="bernoulli",tuneGrid = rf_grid)
  
  #get performance
  preds<-predict(rf_model,holdout,type="prob")[,1]
  predObj<-ROCR::prediction(preds,holdout$stable_sg)
  perf<-ROCR::performance(predObj,"auc")
  rf_roc[i]<-perf@y.values[[1]]
  preds<-predict(rf_model,holdout) %>% factor(levels=c("NoSG","SG"))
  rf_rates<-table(preds ,holdout$stable_sg)/colSums(table(preds,holdout$stable_sg))
  rf_spec[i]<-rf_rates[1,1]
  rf_sens[i]<-rf_rates[2,2]
  print(rf_rates)
  
  preds<-predict(svm_model,holdout,type="prob")[,1]
  predObj<-ROCR::prediction(preds,holdout$stable_sg)
  perf<-ROCR::performance(predObj,"auc")
  svm_roc[i]<-perf@y.values[[1]]
  preds<-predict(svm_model,holdout) %>% factor(levels=c("NoSG","SG"))
  svm_rates<-table(preds,holdout$stable_sg)/colSums(table(preds,holdout$stable_sg))
  svm_spec[i]<-svm_rates[1,1]
  svm_sens[i]<-svm_rates[2,2]
  
    preds<-predict(gbm_model,holdout,type="prob")[,1]
  predObj<-ROCR::prediction(preds,holdout$stable_sg)
  perf<-ROCR::performance(predObj,"auc")
  gbm_roc[i]<-perf@y.values[[1]]
  preds<-predict(gbm_model,holdout) %>% factor(levels=c("NoSG","SG"))
  gbm_rates<-table(preds,holdout$stable_sg)/colSums(table(preds,holdout$stable_sg))
  gbm_spec[i]<-gbm_rates[1,1]
  gbm_sens[i]<-gbm_rates[2,2]
}
print(1-svm_roc)
mean(1-svm_roc)
print(1-rf_roc)
mean(1-rf_roc)
print(1-gbm_roc)
mean(1-gbm_roc)

svm_sens[is.na(svm_sens)]<-0
rf_sens[is.na(rf_sens)]<-0
gbm_sens[is.na(gbm_sens)]<-0

svm_spec[is.na(svm_spec)]<-0
rf_spec[is.na(rf_spec)]<-0
gbm_spec[is.na(gbm_spec)]<-0

mean(svm_sens)
mean(svm_spec)

mean(rf_sens)
mean(rf_spec)

mean(gbm_sens)
mean(gbm_spec)


```
